requirements.txt
pycairo
Pillow